app/requirements.txt
alembic==1.4.3
MarkupSafe==2.0.0
FastAPI-SQLAlchemy==0.2.1
pydantic==2.5.2
python-dotenv==0.21.0
sqlalchemy-json==0.4.0
fastapi==0.105.0
gunicorn==20.0.4
uvicorn==0.12.3
requests==2.25.0
python-jose==3.2.0
bcrypt==3.2.0
passlib==1.7.4
#pyodbc==4.0.32
#psycopg2==2.8.6
email-validator==1.1.2
python-multipart==0.0.5
password-validator==1.0
pydantic-settings==2.1.0
